third task6

+++ src/ru/progwards/java1/lessons/classes/ComplexNum.java
@@ -1,8 +1,8 @@
 package ru.progwards.java1.lessons.classes;
 
 public class ComplexNum {
-    double a;
-    double b;
+    int a;
+    int b;
     public ComplexNum(int a, int b){
         this.a = a;
         this.b = b;
